requirements.txt
python-telegram-bot==21.10
APScheduler==3.10.4
pytz
